requirements.txt
selenium
Pillow
django-tables2
djangorestframework
django-filter
asn1crypto==0.24.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.7
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
idna==2.7
jsonschema==2.6.0
paramiko==2.6.0
psycopg2==2.8.3
pycparser==2.19
PyNaCl==1.3.0
python-dotenv==0.10.3
pytz==2019.1
PyYAML==4.2b1
requests==2.20.1
six==1.12.0
sqlparse==0.3.0
texttable==0.9.1
urllib3==1.24.3
websocket-client==0.56.0
django-webpack-loader==0.6.0
django-extensions==2.2.1
pydotplus==2.0.2
ipython==7.7.0
Werkzeug==0.15.5
factory-boy==2.12.0
django-debug-toolbar==2.0
WeasyPrint==50
channels==2.3.1
channels-redis==2.4.1
aioredis==1.3
